Consolidate checks for POST and GET requests
