+++ internal/src/org/labkey/api/action/GWTServiceAction.java
@@ -40,7 +40,7 @@ public abstract class GWTServiceAction extends BaseViewAction<Object>
     public ModelAndView handleRequest(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse)
     {
         BaseRemoteService service = createService();
-        if (!"POST".equalsIgnoreCase(httpServletRequest.getMethod()))
+        if (!isPost())
         {
             // GWT service requests must be POSTs
             httpServletResponse.setStatus(HttpServletResponse.SC_METHOD_NOT_ALLOWED);
